requirements.txt
django==3.1.6

django-cors-headers==3.5.0
djangorestframework==3.12.2
djoser==2.0.3

pillow==8.0.1
django-cleanup==5.1.0

whitenoise==5.2.0
python-decouple==3.3
gunicorn==20.0.4

mixer==6.1.3
pytest==6.1.2
pytest-cov==2.10.1
pytest-django==4.1.0
dj-inmemorystorage==2.1.0
